+++ tests/test_Signal.js
@@ -488,6 +488,7 @@ tests.test_Signal = function (t) {
     signal(hasSignals, 'signalTwo');
     t.is(testObj.countOne, 0, 'disconnectAllTo obj+str');
     t.is(testObj.countTwo, 1, 'disconnectAllTo obj+str');
+    disconnectAll(hasSignals);
 
     var has__Connect = {
           count: 0,
@@ -523,6 +524,7 @@ tests.test_Signal = function (t) {
     });
     signal(events, "test");
     t.is(events.failed, undefined, 'disconnected slots do not fire');
+    disconnectAll(events);
 
     var sink = {f: function (ev) { this.ev = ev; }};
     var src = {};
@@ -530,6 +532,7 @@ tests.test_Signal = function (t) {
     connect(src, 'signal', sink.f);
     signal(src, 'signal', 'worked');
     t.is(sink.ev, 'worked', 'custom signal does not re-bind methods');
+    disconnectAll(src);
 
     var lateObj = { fun: function() { this.value = 1; } };
     connect(src, 'signal', lateObj, "fun");
@@ -537,4 +540,45 @@ tests.test_Signal = function (t) {
     lateObj.fun = function() { this.value = 2; };
     signal(src, 'signal');
     t.is(lateObj.value, 2, 'connect uses late function binding');
+    disconnectAll(src);
+
+    // Test connectOnce normal scenario
+    var sink = {f: function (ev) { this.ev = ev; }};
+    var src = {};
+    bindMethods(sink);
+    connectOnce(src, 'signal', sink.f);
+    signal(src, 'signal', 'first');
+    t.is(sink.ev, 'first', 'connectOnce propagates first signal');
+    signal(src, 'signal', 'second');
+    t.is(sink.ev, 'first', 'connectOnce disconnects on first signal');
+    disconnectAll(src);
+
+    // Test connectOnce with exception in handler
+    var sink = {f: function (ev) { this.ev = ev; throw new Error("whatever"); }};
+    var src = {};
+    bindMethods(sink);
+    connectOnce(src, 'signal', sink.f);
+    try {
+        signal(src, 'signal', 'first');
+        t.ok(false, 'An exception was not raised in signal handler');
+    } catch (e) {
+        t.ok(true, 'An exception was raised in signal handler');
+    }
+    try {
+        signal(src, 'signal', 'second');
+        t.is(sink.ev, 'first', 'connectOnce disconnected in spite of exception');
+    } catch (e) {
+        t.ok(false, 'connectOnce failed to disconnect due to exception');
+    }
+    disconnectAll(src);
+
+    // Test connectOnce with disconnect before first signal
+    var sink = {f: function (ev) { this.ev = ev; }, ev: "dummy" };
+    var src = {};
+    bindMethods(sink);
+    var ident = connectOnce(src, 'signal', sink.f);
+    disconnect(ident);
+    signal(src, 'signal', 'first');
+    t.is(sink.ev, 'dummy', 'connectOnce was disconnected properly');
+
 };